Major architectural updates (mainly with Docker and Kubernetes management)

+++ app/systems/celery/worker.py
@@ -1,40 +1,33 @@
-from django.conf import settings
-
-from utility.data import ensure_list
-
 import os
 import signal
 import threading
 import time
+
 import redis
+from django.conf import settings
+from utility.data import ensure_list
 
 
 def start_worker_manager(app):
-    return WorkerManager(app)
+    return WorkerManager(app, app.control.inspect())
 
 
-class RedisConnectionMixin(object):
-
+class RedisConnectionMixin:
     def connection(self):
-        if not getattr(self, '_connection', None):
+        if not getattr(self, "_connection", None):
             if settings.CELERY_BROKER_URL:
-                self._connection = redis.from_url(
-                    settings.CELERY_BROKER_URL,
-                    encoding = 'utf-8',
-                    decode_responses = True
-                )
+                self._connection = redis.from_url(settings.CELERY_BROKER_URL, encoding="utf-8", decode_responses=True)
             else:
                 self._connection = None
         return self._connection
 
-
     def get_queues(self, app, worker_name):
-        worker_queues = app.control.inspect().active_queues()
+        worker_queues = self.inspector.active_queues()
         if worker_queues and worker_name in worker_queues:
-            return [ queue_info['name'] for queue_info in worker_queues[worker_name] ]
+            return [queue_info["name"] for queue_info in worker_queues[worker_name]]
         return []
 
-    def check_queues(self, queue_names, app = None, worker_name = None):
+    def check_queues(self, queue_names, app=None, worker_name=None):
         # Find tasks yet to be executed
         for queue_name in ensure_list(queue_names):
             count = self.connection().llen(queue_name)
@@ -42,15 +35,13 @@ class RedisConnectionMixin(object):
                 return True
 
         if app and worker_name:
-            inspector = app.control.inspect()
-
             # Find tasks reserved to be executed
-            reserved = inspector.reserved()
+            reserved = self.inspector.reserved()
             if reserved and reserved.get(worker_name, None):
                 return True
 
             # Find tasks being executed
-            active = inspector.active()
+            active = self.inspector.active()
             if active and active.get(worker_name, None):
                 return True
 
@@ -58,15 +49,14 @@ class RedisConnectionMixin(object):
 
 
 class WorkerManager(RedisConnectionMixin, threading.Thread):
-
-    def __init__(self, app):
+    def __init__(self, app, inspector):
         super().__init__()
         self.app = app
+        self.inspector = inspector
 
         self.daemon = True
         self.stop_signal = threading.Event()
         self.start()
-
 
     def run(self):
         if not self.connection():
@@ -77,11 +67,11 @@ class WorkerManager(RedisConnectionMixin, threading.Thread):
 
             while not self.terminated:
                 if settings.WORKER_TIMEOUT > 0:
-                    worker_name = os.environ.get('ZIMAGI_CELERY_NAME', None)
-                    worker_queues = self.get_queues(self.app, worker_name)
+                    worker_name = os.environ.get("ZIMAGI_CELERY_NAME", None)
+                    if worker_name:
+                        worker_queues = self.get_queues(self.app, worker_name)
 
-                    if worker_name and worker_queues:
-                        if not self.check_queues(worker_queues, app = self.app, worker_name = worker_name):
+                        if worker_queues and not self.check_queues(worker_queues, app=self.app, worker_name=worker_name):
                             if (current_time - start_time) > settings.WORKER_TIMEOUT:
                                 os.kill(os.getpid(), signal.SIGTERM)
                                 break
@@ -93,8 +83,7 @@ class WorkerManager(RedisConnectionMixin, threading.Thread):
         finally:
             self.connection().close()
 
-
-    def terminate(self, timeout = None):
+    def terminate(self, timeout=None):
         self.stop_signal.set()
         super().join(timeout)
